refactor: move `sse test` to `tests` directory
